+++ src/.env
@@ -3,3 +3,4 @@
 #SERVER_ADDR=
 #REDIS_ADDR=
 #POSTGRES_ADDR=
+CORS_ALLOWED_METHODS="GET OPTIONS"
