+++ src/main.rs
@@ -1,83 +1,84 @@
-use futures::{stream::Stream, Async};
 use ragequit::{
-    any_of, config, error,
-    stream_manager::StreamManager,
-    timeline,
-    user::{Filter::*, User},
-    ws,
+    any_of, config,
+    parse_client_request::{sse, user, ws},
+    redis_to_client_stream,
+    redis_to_client_stream::ClientAgent,
 };
+
 use warp::{ws::Ws2, Filter as WarpFilter};
 
 fn main() {
     config::logging_and_env();
-    let stream_manager_sse = StreamManager::new();
-    let stream_manager_ws = stream_manager_sse.clone();
+    let client_agent_sse = ClientAgent::blank();
+    let client_agent_ws = client_agent_sse.clone_with_shared_receiver();
 
     // Server Sent Events
+    //
+    // For SSE, the API requires users to use different endpoints, so we first filter based on
+    // the endpoint.  Using that endpoint determine the `timeline` the user is requesting,
+    // the scope for that `timeline`, and authenticate the `User` if they provided a token.
     let sse_routes = any_of!(
         // GET /api/v1/streaming/user/notification                     [private; notification filter]
-        timeline::user_notifications(),
+        sse::Request::user_notifications(),
         // GET /api/v1/streaming/user                                  [private; language filter]
-        timeline::user(),
+        sse::Request::user(),
         // GET /api/v1/streaming/public/local?only_media=true          [public; language filter]
-        timeline::public_local_media(),
+        sse::Request::public_local_media(),
         // GET /api/v1/streaming/public?only_media=true                [public; language filter]
-        timeline::public_media(),
+        sse::Request::public_media(),
         // GET /api/v1/streaming/public/local                          [public; language filter]
-        timeline::public_local(),
+        sse::Request::public_local(),
         // GET /api/v1/streaming/public                                [public; language filter]
-        timeline::public(),
+        sse::Request::public(),
         // GET /api/v1/streaming/direct                                [private; *no* filter]
-        timeline::direct(),
+        sse::Request::direct(),
         // GET /api/v1/streaming/hashtag?tag=:hashtag                  [public; no filter]
-        timeline::hashtag(),
+        sse::Request::hashtag(),
         // GET /api/v1/streaming/hashtag/local?tag=:hashtag            [public; no filter]
-        timeline::hashtag_local(),
+        sse::Request::hashtag_local(),
         // GET /api/v1/streaming/list?list=:list_id                    [private; no filter]
-        timeline::list()
+        sse::Request::list()
     )
     .untuple_one()
     .and(warp::sse())
-    .map(move |timeline: String, user: User, sse: warp::sse::Sse| {
-        let mut stream_manager = stream_manager_sse.manage_new_timeline(&timeline, user);
-        let event_stream = tokio::timer::Interval::new(
-            std::time::Instant::now(),
-            std::time::Duration::from_millis(100),
-        )
-        .filter_map(move |_| match stream_manager.poll() {
-            Ok(Async::Ready(Some(json_value))) => Some((
-                warp::sse::event(json_value["event"].clone().to_string()),
-                warp::sse::data(json_value["payload"].clone()),
-            )),
-            _ => None,
-        });
-        sse.reply(warp::sse::keep(event_stream, None))
-    })
+    .map(
+        move |timeline: String, user: user::User, sse_connection_to_client: warp::sse::Sse| {
+            // Create a new ClientAgent
+            let mut client_agent = client_agent_sse.clone_with_shared_receiver();
+            // Assign that agent to generate a stream of updates for the user/timeline pair
+            client_agent.init_for_user(&timeline, user);
+            // send the updates through the SSE connection
+            redis_to_client_stream::send_updates_to_sse(client_agent, sse_connection_to_client)
+        },
+    )
     .with(warp::reply::with::header("Connection", "keep-alive"))
-    .recover(error::handle_errors);
+    .recover(config::handle_errors);
 
     // WebSocket
-    let websocket_routes = ws::websocket_routes()
-        .and_then(move |mut user: User, q: ws::Query, ws: Ws2| {
+    //
+    // For WS, the API specifies a single endpoint, so we extract the User/timeline pair
+    // directy from the query
+    let websocket_routes = ws::extract_user_and_query()
+        .and_then(move |mut user: user::User, q: ws::Query, ws: Ws2| {
+            let token = user.access_token.clone();
             let read_scope = user.scopes.clone();
+
             let timeline = match q.stream.as_ref() {
                 // Public endpoints:
                 tl @ "public" | tl @ "public:local" if q.media => format!("{}:media", tl),
                 tl @ "public:media" | tl @ "public:local:media" => tl.to_string(),
                 tl @ "public" | tl @ "public:local" => tl.to_string(),
                 // Hashtag endpoints:
-                // TODO: handle missing query
                 tl @ "hashtag" | tl @ "hashtag:local" => format!("{}:{}", tl, q.hashtag),
                 // Private endpoints: User
                 "user" if user.logged_in && (read_scope.all || read_scope.statuses) => {
                     format!("{}", user.id)
                 }
                 "user:notification" if user.logged_in && (read_scope.all || read_scope.notify) => {
-                    user = user.set_filter(Notification);
+                    user = user.set_filter(user::Filter::Notification);
                     format!("{}", user.id)
                 }
                 // List endpoint:
-                // TODO: handle missing query
                 "list" if user.owns_list(q.list) && (read_scope.all || read_scope.lists) => {
                     format!("list:{}", q.list)
                 }
@@ -92,11 +93,17 @@ fn main() {
                 // Other endpoints don't exist:
                 _ => return Err(warp::reject::custom("Error: Nonexistent WebSocket query")),
             };
-            let token = user.access_token.clone();
-            let stream_manager = stream_manager_ws.manage_new_timeline(&timeline, user);
 
+            // Create a new ClientAgent
+            let mut client_agent = client_agent_ws.clone_with_shared_receiver();
+            // Assign that agent to generate a stream of updates for the user/timeline pair
+            client_agent.init_for_user(&timeline, user);
+            // send the updates through the WS connection (along with the User's access_token
+            // which is sent for security)
             Ok((
-                ws.on_upgrade(move |socket| ws::send_replies(socket, stream_manager)),
+                ws.on_upgrade(move |socket| {
+                    redis_to_client_stream::send_updates_to_ws(socket, client_agent)
+                }),
                 token,
             ))
         })
