+++ src/redis_to_client_stream/receiver.rs
@@ -1,16 +1,13 @@
-//! Interface with Redis and stream the results to the `StreamManager`
-//! There is only one `Receiver`, which suggests that it's name is bad.
-//!
-//! **TODO**: Consider changing the name.  Maybe RedisConnectionPool?
-//! There are many AsyncReadableStreams, though.  How do they fit in?
-//! Figure this out ASAP.
-//! A new one is created every time the Receiver is polled
-use crate::{config, pubsub_cmd, redis_cmd};
+//! Receives data from Redis, sorts it by `ClientAgent`, and stores it until
+//! polled by the correct `ClientAgent`.  Also manages sububscriptions and
+//! unsubscriptions to/from Redis.
+use super::redis_cmd;
+use crate::{config, pubsub_cmd};
 use futures::{Async, Poll};
 use log::info;
 use regex::Regex;
 use serde_json::Value;
-use std::{collections, io::Read, io::Write, net, time};
+use std::{collections, env, io::Read, io::Write, net, time};
 use tokio::io::{AsyncRead, Error};
 use uuid::Uuid;
 
@@ -19,7 +16,8 @@ use uuid::Uuid;
 pub struct Receiver {
     pubsub_connection: net::TcpStream,
     secondary_redis_connection: net::TcpStream,
-    tl: String,
+    redis_polled_at: time::Instant,
+    timeline: String,
     manager_id: Uuid,
     msg_queues: collections::HashMap<Uuid, MsgQueue>,
     clients_per_timeline: collections::HashMap<String, i32>,
@@ -33,7 +31,8 @@ impl Receiver {
         Self {
             pubsub_connection,
             secondary_redis_connection,
-            tl: String::new(),
+            redis_polled_at: time::Instant::now(),
+            timeline: String::new(),
             manager_id: Uuid::default(),
             msg_queues: collections::HashMap::new(),
             clients_per_timeline: collections::HashMap::new(),
@@ -43,60 +42,60 @@ impl Receiver {
     /// Assigns the `Receiver` a new timeline to monitor and runs other
     /// first-time setup.
     ///
-    /// Importantly, this method calls `subscribe_or_unsubscribe_as_needed`,
+    /// Note: this method calls `subscribe_or_unsubscribe_as_needed`,
     /// so Redis PubSub subscriptions are only updated when a new timeline
     /// comes under management for the first time.
     pub fn manage_new_timeline(&mut self, manager_id: Uuid, timeline: &str) {
         self.manager_id = manager_id;
-        self.tl = timeline.to_string();
-        let old_value = self
-            .msg_queues
+        self.timeline = timeline.to_string();
+        self.msg_queues
             .insert(self.manager_id, MsgQueue::new(timeline));
-        // Consider removing/refactoring
-        if let Some(value) = old_value {
-            eprintln!(
-                "Data was overwritten when it shouldn't have been.  Old data was: {:#?}",
-                value
-            );
-        }
         self.subscribe_or_unsubscribe_as_needed(timeline);
     }
 
     /// Set the `Receiver`'s manager_id and target_timeline fields to the approprate
     /// value to be polled by the current `StreamManager`.
     pub fn configure_for_polling(&mut self, manager_id: Uuid, timeline: &str) {
-        if &manager_id != &self.manager_id {
-            //println!("New Manager: {}", &manager_id);
-        }
         self.manager_id = manager_id;
-        self.tl = timeline.to_string();
+        self.timeline = timeline.to_string();
     }
 
     /// Drop any PubSub subscriptions that don't have active clients and check
     /// that there's a subscription to the current one.  If there isn't, then
     /// subscribe to it.
-    fn subscribe_or_unsubscribe_as_needed(&mut self, tl: &str) {
+    fn subscribe_or_unsubscribe_as_needed(&mut self, timeline: &str) {
         let mut timelines_to_modify = Vec::new();
-        timelines_to_modify.push((tl.to_owned(), 1));
+        struct Change {
+            timeline: String,
+            change_in_subscriber_number: i32,
+        }
+
+        timelines_to_modify.push(Change {
+            timeline: timeline.to_owned(),
+            change_in_subscriber_number: 1,
+        });
 
         // Keep only message queues that have been polled recently
         self.msg_queues.retain(|_id, msg_queue| {
             if msg_queue.last_polled_at.elapsed() < time::Duration::from_secs(30) {
                 true
             } else {
-                let timeline = msg_queue.redis_channel.clone();
-                timelines_to_modify.push((timeline, -1));
+                let timeline = &msg_queue.redis_channel;
+                timelines_to_modify.push(Change {
+                    timeline: timeline.to_owned(),
+                    change_in_subscriber_number: -1,
+                });
                 false
             }
         });
 
         // Record the lower number of clients subscribed to that channel
-        for (timeline, numerical_change) in timelines_to_modify {
+        for change in timelines_to_modify {
             let mut need_to_subscribe = false;
             let count_of_subscribed_clients = self
                 .clients_per_timeline
-                .entry(timeline.to_owned())
-                .and_modify(|n| *n += numerical_change)
+                .entry(change.timeline.clone())
+                .and_modify(|n| *n += change.change_in_subscriber_number)
                 .or_insert_with(|| {
                     need_to_subscribe = true;
                     1
@@ -104,11 +103,38 @@ impl Receiver {
             // If no clients, unsubscribe from the channel
             if *count_of_subscribed_clients <= 0 {
                 info!("Sent unsubscribe command");
-                pubsub_cmd!("unsubscribe", self, timeline.clone());
+                pubsub_cmd!("unsubscribe", self, change.timeline.clone());
             }
             if need_to_subscribe {
                 info!("Sent subscribe command");
-                pubsub_cmd!("subscribe", self, timeline.clone());
+                pubsub_cmd!("subscribe", self, change.timeline.clone());
+            }
+        }
+    }
+
+    /// Polls Redis for any new messages and adds them to the `MsgQueue` for
+    /// the appropriate `ClientAgent`.
+    fn poll_redis(&mut self) {
+        let mut buffer = vec![0u8; 3000];
+        // Add any incoming messages to the back of the relevant `msg_queues`
+        // NOTE: This could be more/other than the `msg_queue` currently being polled
+        let mut async_stream = AsyncReadableStream::new(&mut self.pubsub_connection);
+        if let Async::Ready(num_bytes_read) = async_stream.poll_read(&mut buffer).unwrap() {
+            let raw_redis_response = &String::from_utf8_lossy(&buffer[..num_bytes_read]);
+            // capture everything between `{` and `}` as potential JSON
+            let json_regex = Regex::new(r"(?P<json>\{.*\})").expect("Hard-coded");
+            // capture the timeline so we know which queues to add it to
+            let timeline_regex = Regex::new(r"timeline:(?P<timeline>.*?)\r").expect("Hard-codded");
+            if let Some(result) = json_regex.captures(raw_redis_response) {
+                let timeline =
+                    timeline_regex.captures(raw_redis_response).unwrap()["timeline"].to_string();
+
+                let msg: Value = serde_json::from_str(&result["json"].to_string().clone()).unwrap();
+                for msg_queue in self.msg_queues.values_mut() {
+                    if msg_queue.redis_channel == timeline {
+                        msg_queue.messages.push_back(msg.clone());
+                    }
+                }
             }
         }
     }
@@ -128,46 +154,40 @@ impl Receiver {
         }
     }
 }
+
 impl Default for Receiver {
     fn default() -> Self {
         Receiver::new()
     }
 }
 
+/// The stream that the ClientAgent polls to learn about new messages.
 impl futures::stream::Stream for Receiver {
     type Item = Value;
     type Error = Error;
 
+    /// Returns the oldest message in the `ClientAgent`'s queue (if any).
+    ///
+    /// Note: This method does **not** poll Redis every time, because polling
+    /// Redis is signifiantly more time consuming that simply returning the
+    /// message already in a queue.  Thus, we only poll Redis if it has not
+    /// been polled lately.
     fn poll(&mut self) -> Poll<Option<Value>, Self::Error> {
-        let mut buffer = vec![0u8; 3000];
-        let timeline = self.tl.clone();
+        let timeline = self.timeline.clone();
+
+        let redis_poll_interval = env::var("REDIS_POLL_INTERVAL")
+            .map(|s| s.parse().expect("Valid config"))
+            .unwrap_or(config::DEFAULT_REDIS_POLL_INTERVAL);
+
+        if self.redis_polled_at.elapsed() > time::Duration::from_millis(redis_poll_interval) {
+            self.poll_redis();
+            self.redis_polled_at = time::Instant::now();
+        }
 
         // Record current time as last polled time
         self.msg_queues
             .entry(self.manager_id)
             .and_modify(|msg_queue| msg_queue.last_polled_at = time::Instant::now());
-
-        // Add any incomming messages to the back of the relevant `msg_queues`
-        // NOTE: This could be more/other than the `msg_queue` currently being polled
-        let mut async_stream = AsyncReadableStream::new(&mut self.pubsub_connection);
-        if let Async::Ready(num_bytes_read) = async_stream.poll_read(&mut buffer)? {
-            let raw_redis_response = &String::from_utf8_lossy(&buffer[..num_bytes_read]);
-            // capture everything between `{` and `}` as potential JSON
-            let json_regex = Regex::new(r"(?P<json>\{.*\})").expect("Hard-coded");
-            // capture the timeline so we know which queues to add it to
-            let timeline_regex = Regex::new(r"timeline:(?P<timeline>.*?)\r").expect("Hard-codded");
-            if let Some(result) = json_regex.captures(raw_redis_response) {
-                let timeline =
-                    timeline_regex.captures(raw_redis_response).unwrap()["timeline"].to_string();
-
-                let msg: Value = serde_json::from_str(&result["json"].to_string().clone())?;
-                for msg_queue in self.msg_queues.values_mut() {
-                    if msg_queue.redis_channel == timeline {
-                        msg_queue.messages.push_back(msg.clone());
-                    }
-                }
-            }
-        }
 
         // If the `msg_queue` being polled has any new messages, return the first (oldest) one
         match self
@@ -188,7 +208,7 @@ impl futures::stream::Stream for Receiver {
 
 impl Drop for Receiver {
     fn drop(&mut self) {
-        pubsub_cmd!("unsubscribe", self, self.tl.clone());
+        pubsub_cmd!("unsubscribe", self, self.timeline.clone());
     }
 }
